update: zig version 0.15.1 (#97)
* update everything up to libvaxis api changes

* replace sleep api

* libvaxis changes

* temp u16 fix

* correctly allocate and deallocate

* update libvaxis

* update github actions

+++ src/Cache.zig
@@ -25,7 +25,7 @@ map: std.AutoHashMap(Key, *Node),
 head: ?*Node,
 tail: ?*Node,
 config: *Config,
-lru_size: usize,
+lru_size: u16,
 vx: vaxis.Vaxis,
 tty: *const vaxis.Tty,
